added authorization to airtable integration and made a common folder for actions
