+++ apps/web/app/(app)/environments/[environmentId]/integrations/google-sheets/components/Home.tsx
@@ -1,6 +1,5 @@
 "use client";
-
-import { deleteIntegrationAction } from "@/app/(app)/environments/[environmentId]/integrations/google-sheets/actions";
+import { deleteIntegrationAction } from "@/app/(app)/environments/[environmentId]/integrations/actions";
 import { timeSince } from "@formbricks/lib/time";
 import { TEnvironment } from "@formbricks/types/environment";
 import {
